+++ Changelog.md
@@ -5,6 +5,9 @@ This file exists because sometimes I wait to commit my changes until I've done a
 - Remove axios request timeout
 - Add invite command
 - Add `/users list` subcommand
+- Add context menu command for regenerating a response
+- Add "Try again" button on error messages
+- Add config option for showing "Regenerate" button on responses
 
 ## 2023-03-18
 - Add timestamps and colour to logs